fix: resolve atoms build error
