+++ packages/app-store/hitpay/components/constants.ts
@@ -1,3 +1,51 @@
+export const paymentOptions = [
+  {
+    label: "on_booking_option",
+    value: "ON_BOOKING",
+  },
+];
+
+export const supportedPaymentMethods = [
+  "paynow",
+  "grabpay",
+  "shopeepay",
+  "zip",
+  "fpx",
+  "visa",
+  "master",
+  "american_express",
+  "alipay",
+  "atome",
+  "unionpay",
+  "gcash",
+  "qrph",
+  "billease",
+  "seveneleven",
+  "cebuana",
+  "palawa",
+  "ovo",
+  "gopay",
+  "linkaja",
+  "dana",
+  "kredivo",
+  "akulakupaylater",
+  "akulaku",
+  "indomaret",
+  "alfamart",
+  "bri",
+  "cimb",
+  "sbpl",
+  "payid",
+  "qris",
+  "bdo",
+  "bpi",
+  "duitnow",
+  "wechatpay",
+  "qr_promptpay",
+  "line_pay",
+  "truemoney_pay",
+];
+
 export const currencyOptions = [
   { label: "United States dollar (USD)", value: "usd" },
   { label: "Singapore dollar (SGD)", value: "sgd" },
