+++ packages/app-store/hitpay/lib/PaymentService.ts
@@ -75,6 +75,7 @@ export class PaymentService implements IAbstractPaymentService {
         redirect_url: redirectUri,
         webhook: webhookUri,
         channel: "api_cal",
+        purpose: booking.title,
       };
 
       const response = await axios.post(requestUrl, qs.stringify(formData), {
